# Improve performance for map and union columns