Before:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 2.939 s | 0.4963 s | 0.3283 s |               385972 |

After change to avx operation in union column offset list:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 2.306 s | 0.5179 s | 0.3425 s |               494721 |

Remove sorting in map add and insert since its already sorted:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 2.223 s | 0.5163 s | 0.3415 s |               513782 |